requirements.txt
django >= 1.7, < 1.8
django-admin-bootstrapped
django-bootstrap3
django-registration-redux
django-extensions
django-import-export
